requirements.txt
python-telegram-bot>=21.10,<22.0
APScheduler>=3.10,<4.0
motor>=3.7,<4.0
python-dotenv>=1.0,<2.0
Pillow>=10.4,<12.0
nudenet>=3.4,<4.0
